# swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/eventqueue": {
            "put": {
                "description": "Update by json EventQueue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventQueue"
                ],
                "summary": "Update EventQueue",
                "parameters": [
                    {
                        "description": "Update EventQueue",
                        "name": "EventQueue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EventQueue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "create EventQueue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventQueue"
                ],
                "summary": "Create EventQueue",
                "parameters": [
                    {
                        "description": "create EventQueue",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EventQueue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/eventqueue/{id}": {
            "get": {
                "description": "Get EventQueue by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventQueue"
                ],
                "summary": "Get all EventQueue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EventQueue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventQueue"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete EventQueue",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventQueue"
                ],
                "summary": "Get all EventQueue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EventQueue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/eventqueues/search": {
            "get": {
                "description": "get top 100 records EventQueue",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventQueue"
                ],
                "summary": "Get all EventQueue",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Starting row 0--\u003efirst record",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Row limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventQueue"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.EventQueue": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "createdAt",
                    "type": "string"
                },
                "eventData": {
                    "description": "eventData",
                    "type": "string"
                },
                "eventTypeId": {
                    "description": "eventTypeId",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "isActive": {
                    "description": "isActive",
                    "type": "boolean"
                },
                "message": {
                    "description": "Message",
                    "type": "string"
                },
                "retryCount": {
                    "description": "retryCount",
                    "type": "integer"
                },
                "startAt": {
                    "description": "startAt",
                    "type": "string"
                },
                "status": {
                    "description": "status",
                    "type": "integer"
                },
                "updatedAt": {
                    "description": "updatedAt",
                    "type": "string"
                }
            }
        },
...
    }
}